Script and Results
Script for brute forcing via SSH. The text file included shows the results of the script. rockyou.txt was not included due to size constraints on the document. For testing purposes the known password of the metasploitable machine was added at the beginning of the rockyou.txt. This script was written merely as a project to connect to a remote machine and attempt to login using brute force and a txt file.